doesn't compile after declawing because of an autoboxing usage...
git-svn-id: https://svn.java.net/svn/glazedlists~svn/trunk@1203 696c4f5c-992c-4387-df9e-f6c076136008

+++ test/ca/odell/glazedlists/ThresholdListTest.java
@@ -1202,7 +1202,7 @@ public class ThresholdListTest extends TestCase {
      */
     private class IntegerAsStringEvaluator implements ThresholdList.Evaluator {
         public int evaluate(Object object) {
-            return Integer.valueOf((String)object);
+            return Integer.valueOf((String)object).intValue();
         }
     }
